# docs/swagger.yaml
basePath: /api/sf/v1
definitions:
  db.Pagination:
    properties:
      currentPage:
        type: integer
      size:
        type: integer
      skipPagination:
        type: boolean
      sort:
        type: string
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
...
  errs.ErrorType:
    enum:
    - OTHER
    - INVALID
    - IO
    - EXIST
    - NOT_EXIST
    - PRIVATE
    - INTERNAL
    - BROKEN_LINK
    - DATABASE
    - VALIDATION
    - UNANTICIPATED
    - INVALID_REQUEST
    - SMTP_ERROR
    - EMAIL_ERROR
    - STORAGE
    - INSUFFICIENT_ACCESS
    - CUSTOMER_DUPLICATE_TRANSACTION
    type: string
    x-enum-comments:
      BROKEN_LINK: Link target does not exist.
      CUSTOMER_DUPLICATE_TRANSACTION: Duplicate Transaction
      DATABASE: Error from database.
      EMAILERROR: SMTP Error
      EXIST: Item already exists.
      INSUFFICIENT_ACCESS: Insufficient Access
      INTERNAL: Internal error or inconsistency.
      INVALID: Invalid operation for this type of item.
      INVALID_REQUEST: Invalid Request
      IO: External I/O error such as network failure.
      NOT_EXIST: Item does not exist.
      OTHER: Unclassified error. This value is not printed in the error message.
      PRIVATE: Information withheld.
      SMTPERROR: SMTP Error
      STORAGE: Storage Error
      UNANTICIPATED: Unanticipated error.
      VALIDATION: Input validation error.
    x-enum-descriptions:
    - Unclassified error. This value is not printed in the error message.
    - Invalid operation for this type of item.
    - External I/O error such as network failure.
    - Item already exists.
    - Item does not exist.
    - Information withheld.
    - Internal error or inconsistency.
    - Link target does not exist.
    - Error from database.
    - Input validation error.
    - Unanticipated error.
    - Invalid Request
    - SMTP Error
    - SMTP Error
    - Storage Error
    - Insufficient Access
    - Duplicate Transaction
    x-enum-varnames:
    - OTHER
    - INVALID
    - IO
    - EXIST
    - NOT_EXIST
    - PRIVATE
    - INTERNAL
    - BROKEN_LINK
    - DATABASE
    - VALIDATION
    - UNANTICIPATED
    - INVALID_REQUEST
    - SMTPERROR
    - EMAILERROR
    - STORAGE
    - INSUFFICIENT_ACCESS
    - CUSTOMER_DUPLICATE_TRANSACTION
  errs.XError:
    properties:
      error: {}
      errorCode:
        type: string
      errorType:
        $ref: '#/definitions/errs.ErrorType'
      params:
        items: {}
        type: array
    type: object
...
  requestModel.Category:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
    type: object
...
  requestModel.Inventory:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
      lowStockThreshold:
        type: integer
      productId:
        type: integer
    type: object
...
  requestModel.StockMovementRequest:
    properties:
      adminOverride:
        description: Allow OUT even if stock insufficient
        type: boolean
      changeType:
        description: IN, OUT, ADJUST
        type: string
      notes:
        type: string
      productId:
        type: integer
      quantity:
        type: integer
      reason:
        type: string
    required:
    - changeType
    - productId
    - quantity
    - reason
    type: object
...
  response.DataResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errs.XError'
      message:
        type: string
      pageMetaData:
        $ref: '#/definitions/db.Pagination'
      type:
        $ref: '#/definitions/response.ResponseType'
    type: object
  response.Response:
    properties:
      error:
        $ref: '#/definitions/errs.XError'
      message:
        type: string
      pageMetaData:
        $ref: '#/definitions/db.Pagination'
      type:
        $ref: '#/definitions/response.ResponseType'
    type: object
  response.ResponseType:
    enum:
    - Success
    - Failure
    - Warning
    type: string
    x-enum-varnames:
    - SUCCESS
    - FAILURE
    - WARNING
...
  responseModel.Category:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      productCount:
        description: Count of products in this category
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  responseModel.CategoryAutoComplete:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  responseModel.Inventory:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isLowStock:
        type: boolean
      lowStockThreshold:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/responseModel.Product'
        description: Related data
      productId:
        type: integer
      productName:
        type: string
      productSku:
        type: string
      quantity:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  responseModel.LowStockItem:
    properties:
      categoryName:
        type: string
      currentStock:
        type: integer
      lowStockThreshold:
        type: integer
      productId:
        type: integer
      productName:
        type: string
      productSku:
        type: string
    type: object
...
  responseModel.Product:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/responseModel.Category'
        description: Related data
      categoryId:
        type: integer
      categoryName:
        description: Flattened category name
        type: string
      costPrice:
        type: number
      createdAt:
        type: string
      createdBy:
        type: string
      currentStock:
        description: From inventory
        type: integer
      description:
        type: string
      id:
        type: integer
      inventory:
        $ref: '#/definitions/responseModel.Inventory'
      isActive:
        type: boolean
      isLowStock:
        description: Stock alert flag
        type: boolean
      name:
        type: string
      sellingPrice:
        type: number
      sku:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  responseModel.StockMovementResponse:
    properties:
      changeAmount:
        type: integer
      message:
        type: string
      newStock:
        type: integer
      previousStock:
        type: integer
      productId:
        type: integer
      success:
        type: boolean
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: This is the backend for Stitchfolio.
  title: Stitchfolio-backend API docs
  version: "1.0"
paths:
...
  /category:
    get:
      consumes:
      - application/json
      description: Get all active categories
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get all active categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Saves an instance of Category
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requestModel.Category'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Save Category
      tags:
      - Category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an instance of Category
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get an instance of Category
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get a specific Category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Updates an instance of Category
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requestModel.Category'
      - description: Category id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Category
      tags:
      - Category
  /category/autocomplete:
    get:
      consumes:
      - application/json
      description: Autocomplete for categories
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.CategoryAutoComplete'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Autocomplete for categories
      tags:
      - Category
...
  /inventory:
    get:
      consumes:
      - application/json
      description: Get all inventory records
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Inventory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get all inventories
      tags:
      - Inventory
...
  /inventory/{id}:
    get:
      consumes:
      - application/json
      description: Get an instance of Inventory
      parameters:
      - description: Inventory id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Inventory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get a specific Inventory
      tags:
      - Inventory
  /inventory/{id}/threshold:
    put:
      consumes:
      - application/json
      description: Update the low stock threshold for a product
      parameters:
      - description: inventory
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/requestModel.Inventory'
      - description: Inventory id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update low stock threshold
      tags:
      - Inventory
  /inventory/low-stock:
    get:
      consumes:
      - application/json
      description: Get all items with stock below threshold
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.LowStockItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get low stock items
      tags:
      - Inventory
  /inventory/movement:
    post:
      consumes:
      - application/json
      description: Record a stock IN, OUT, or ADJUST movement
      parameters:
      - description: stock movement
        in: body
        name: movement
        required: true
        schema:
          $ref: '#/definitions/requestModel.StockMovementRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responseModel.StockMovementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Record stock movement
      tags:
      - Inventory
  /inventory/product/{productId}:
    get:
      consumes:
      - application/json
      description: Get inventory for a specific product
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Inventory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get inventory by product ID
      tags:
      - Inventory
...
swagger: "2.0"